Added `about us` and `contact us` URL
Now Clicking "About Us" scrolls down to #aboutus. May be this isn't necessary
Clicking on "Contact Us" opens up "http://feedback.nitdgplug.org/"

Checking if everything works without `baseurl` in "_config.yml"

Everything works without `baseurl` in "_config.yml"

Everything works because `baseurl` is nothing which makes things like
"https://lugnitdgp.github.io/css/main.css" to just "/css/main.css" which is
easier to parse.
If it is really required in future then it has to be reinitialized.
This was done because setting `baseurl` gives an error when running on localhost
so why not just not set it.

Fix for navigating back to home from blog

Due to sudden removal of `baseurl` the home URL was rendered useless. This
updates it to use "/" for the home URL